requirements.txt
pillow=10.2.0